requirements.txt
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0
pytest>=7.3.1
mypy>=1.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
docker-compose>=1.29.2